Модуляция. main.js был разбит на ES2015-модули. Подключён main.js как модуль в index.html
