+++ index.html
@@ -235,8 +235,8 @@
   <!-- Подключение JavaScript файлов -->
   <script src="js/functions.js"></script>
 
-  <!-- Подключение main.js -->
-  <script src="js/main.js"></script>
+  <!-- Подключение main.js как модуль -->
+  <script type="module" src="./js/main.js"></script>
 
 </body>
 </html>
